requirements.txt
streamlit
opencv-python
imutils
dlib
numpy
scikit-image
absl-py
altair
annotated-types
anyio
astunparse
attrs
audioread
blinker
blis
cached-property
cachetools
catalogue
certifi
cffi
charset-normalizer
chex
click
cloudpathlib
colorama
confection
contourpy
ctransformers
cycler
cymem
decorator
distro
dlib
dm-tree
en-core-web-sm
etils
exceptiongroup
ffmpeg
ffmpeg-python
filelock
flatbuffers
flax
fonttools
fsspec
future
gast
gitdb
GitPython
google-auth
google-auth-oauthlib
google-pasta
groq
grpcio
h11
h5py
httpcore
httpx
huggingface-hub
idna
imageio
importlib_resources
imutils
jax
jaxlib
Jinja2
joblib
jsonschema
jsonschema-specifications
keras
kiwisolver
langcodes
lazy-loader
libclang
librosa
llvmlite
Markdown
markdown-it-py
MarkupSafe
matplotlib
mdurl
ml-dtypes
mpmath
msgpack
murmurhash
namex
nest-asyncio
networkx
nltk
numba
numpy
oauthlib
opencv-python
opt-einsum
optax
optree
orbax-checkpoint
packaging
pandas
pillow
pip
platformdirs
pooch
preshed
protobuf
py-cpuinfo
pyarrow
pyasn1
pyasn1-modules
pycparser
pydantic
pydantic-core
pydeck
pydub
Pygments
pyparsing
python-dateutil
pytz
PyYAML
referencing
regex
requests
requests-oauthlib
rich
rpds-py
rsa
safetensors
scikit-image
scikit-learn
scipy
setuptools
six
smart-open
smmap
sniffio
soundfile
soxr
spacy
spacy-legacy
spacy-loggers
SpeechRecognition
srsly
streamlit
sympy
tenacity
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-intel
tensorflow-io-gcs-filesystem
tensorstore
termcolor
textblob
tf-keras
thinc
threadpoolctl
tifffile
tokenizers
toml
toolz
torch
tornado
tqdm
transformers
typer
typing-extensions
tzdata
urllib3
wasabi
watchdog
weasel
Werkzeug
wheel
whisper
whisper-jax
wrapt
zipp
